builtins: fix remaining incorrect MakeDTimestampTZFromDate usages
Auditing the remaining MakeDTimestampTZFromDate usages and found some
interesting items that do not match postgres behaviour, so fixing all
that up here. Note that the MakeDTimestampTZFromDate bug fix actually
fixed comparators compared to previous behaviour.

This in particular fixes types and return results for:

* date +- interval
* interval + date

And, in addition to the above, adds tests for:
* date +- time
* time + date
* comparator operators.

Release note (bug fix, backward-incompatible change): This PR changes
the return type of (date +- interval) and (interval + date) to be
timestamp instead of timestamptz, to be in line with postgres.
Furthermore, it fixes a bug where these calculations would be incorrect
if the current timezone is not UTC.