tree: modify MakeDTimestampTZ callers to use context location
This PR involves my audit of eval.go and builtins.go to make sure all
TimestampTZ operations are performed in the correct timezone. Results
are described in the release note.

Release note (bug fix, sql change): Previously, some TimestampTZ
operations assumed UTC timezone instead of local timezone.

This is different from postgres for a couple of reasons:
* it leads to bugs involving daylight saving in arithmetic,
e.g. with `America/Chicago`, evaluating
'2010-11-06 23:59:00-05'::timestamptz + '1 day'::interval would return
incorrect results as it assumed it was a fixed offset of `-5` instead.
* it is wrong when directly evaluating objects compared to postgres,
e.g. `SELECT '2010-09-28 12:00:00.1-04:00'::timestamptz` should always
assume the context timestamp (and print `2010-09-28 16:00:00.1` for
a ctx with UTC), instead of returning the same `2010-09-28
12:00:00.1-04:00`.

This fixes the behavior such that any timestamptz parsed resolves in the
context timezone, matching postgres. Note this behavior was not present
when stored, as it would convert to unix timestamp automatically, and be
reparsed out as the context timestamp.