geo: initial framework to parse WKT from GEOS
This PR introduces the ability to parse WKT using the GEOS library. It
is the only library I can tell that can decode WKT, so everything will
have to go through that (for now). We're planning to store everything in
EWKB, and everything can be parsed off WKB which is why I've made
that the base type for now.

To avoid a circular dependency between geo/geos <-> geo, we need to
define the parse functions in geomfn/geogfn until we get WKT out of
there.

In future, the geos package will take in geo.Geometry types, transform
them to WKB to be reparsed back into GEOSGeometry for operations.

Release note: None