sql: implement the client_min_messages session var
PostgreSQL supports notices of varying severity to display to the user.
This is achieved through the `client_min_messages` session var. The
strings it accepts is different than the Severity value.

As such, we create a new pgnotice package that has a `DisplaySeverity`
setting that encodes this value. Severities from pgerror will get
translated into a DisplaySeverity and compared to the
`client_min_messages` value before displaying. Furthermore, `DEBUG<num>`
severities will get truncated to just `DEBUG` for notices.

In addition, we have created the `pgnotice` package which
contains the Severity constants as well as moving `pgerror.Noticef` to
`pgnotice.Newf`.

Release note (sql change): Users now have the ability to control which
level of NOTICE output with the session variable `client_min_messages`.