build/deploy: add GEOS libraries to CRDB Docker builds
Now that we have the GEOS libraries being built, it's time we copy them
into the right place in the Docker container such that users can import
geospatial features out of the box.

In the Makefile, use `cp -L` instead of `ln -sf` for the GEOS libraries,
as `cp -L` doesn't follow symlinks inside a the builder container.

I will change the release script to copy the files over (they are not
committed to the codebase) after this merges.

Release note (general change): The Docker container that ships with
CockroachDB now includes the GEOS library needed for geospatial
functionality in `/usr/local/lib/cockroach` (which is the default
location of where the cockroach binary looks for the GEOS libraries).