sql: introduce ContainsAggregateClass for builtins
We formerly expected all overloads of an AggregateClass to only contain
aggregates. However, as is the case in PostGIS, this is not strictly
true. To account for this, add a new class `ContainsAggregateClass`,
which behaves mostly as AggregateClass except that it allows
non-aggregates to be included the the typename, hence showing up in the
correct places in docs and keeping the error messages for pure
AggregateClass functions.

This is demonstrated with ST_MakeLine, which has both an aggregate
version (existing and tested) and the non-aggregate version.

Release note (sql change): Implement the ST_MakeLine(a, b) variant,
which takes in two geometries and returns a line out of both.