sql: implement ALTER TABLE LOCALITY !(REGIONAL BY ROW) -> REGIONAL BY ROW
This change introduces the ability to add a column before doing an alter
primary key for transforming to REGIONAL BY ROW if the crdb_region does
not already exist. We exclusively whitelist the check in AlterPrimaryKey
which allows an add column mutation from SET LOCALITY CHANGES. Since we
do not run this mutation in conjunction with any other schema changes,
it is a safer space to reason about.

The rest of the logic "falls out" fairly easily since the schema changer
seems to be able to support this state. Tests have been added to
simulate cancels during both the ADD COLUMN and ADD PRIMARY KEY step.

Release note (sql change): Implement the ability to ALTER TABLE SET
LOCALITY to REGIONAL BY ROW without specifying AS for global and
regional by table items.